Move linux dependency under build tag
