+++ consumer_config.go
@@ -3,10 +3,8 @@ package taskq
 import (
 	"fmt"
 	"math/rand"
-	"runtime"
 	"time"
 
-	"github.com/capnm/sysinfo"
 	"github.com/vmihailenco/taskq/v3/internal"
 )
 
@@ -316,26 +314,4 @@ func (r *configRoulette) withMoreFetchers(cfg *consumerConfig) *consumerConfig {
 
 func significant(n float64) bool {
 	return n > 0.05
-}
-
-func hasFreeSystemResources() bool {
-	si := sysinfo.Get()
-	free := si.FreeRam + si.BufferRam
-
-	// at least 200MB of RAM is free
-	if free < 2e5 {
-		return false
-	}
-
-	// at least 5% of RAM is free
-	if float64(free)/float64(si.TotalRam) < 0.05 {
-		return false
-	}
-
-	// avg load is not too high
-	if si.Loads[0] > 1.5*float64(runtime.NumCPU()) {
-		return false
-	}
-
-	return true
 }
