+++ sysinfo_other.go
@@ -1,0 +1,7 @@
+// +build !linux
+
+package taskq
+
+func hasFreeSystemResources() bool {
+	return true
+}
